Rearrange QSPanel UI
