+++ packages/SystemUI/res/layout/quick_status_bar_expanded_header.xml
@@ -23,21 +23,14 @@
     android:layout_height="@dimen/status_bar_header_height"
     android:layout_gravity="@integer/notification_panel_layout_gravity"
     android:baselineAligned="false"
+	android:elevation="4dp"
     android:clickable="false"
     android:clipChildren="false"
     android:clipToPadding="false"
     android:paddingTop="0dp">
-	
-	<View
-        android:id="@+id/qs_top_background"
-        android:layout_width="match_parent"
-        android:layout_height="82dp"
-        android:layout_alignParentTop="true"
-        android:background="@drawable/qs_background_top"
-		android:elevation="0dp"/>
 
     <include layout="@layout/quick_status_bar_header_system_icons" />
-	
+
 	<com.android.systemui.qs.QuickQSPanel
         android:id="@+id/quick_qs_panel"
         android:layout_width="match_parent"
@@ -53,15 +46,6 @@
         android:padding="16dp"
         android:focusable="true"
         android:importantForAccessibility="yes" />
-		
-	<ImageView
-        android:id="@+id/quick_shape"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@android:color/transparent"
-        android:src="@drawable/ic_qs_shape"
-		android:layout_below="@+id/quick_qs_panel"
-		android:visibility="gone"/>
 
     <com.android.systemui.statusbar.AlphaOptimizedImageView
         android:id="@+id/qs_detail_header_progress"
