+++ packages/SystemUI/res/layout/quick_status_bar_header_system_icons.xml
@@ -23,7 +23,8 @@
     android:layout_alignParentEnd="true"
     android:clipChildren="false"
     android:clipToPadding="false"
-    android:gravity="center_vertical">
+    android:gravity="center_vertical"
+	android:background="@drawable/qs_background_top">
 
 
     <com.android.systemui.statusbar.policy.QSClockLeft
