+++ packages/SystemUI/res/layout/status_bar_expanded.xml
@@ -54,8 +54,8 @@
             android:layout_height="match_parent"
             android:layout_gravity="@integer/notification_panel_layout_gravity"
             android:elevation="4dp"
-			android:layout_marginStart="8dp"
-			android:layout_marginEnd="8dp"
+			android:layout_marginStart="@dimen/qs_container_margin"
+			android:layout_marginEnd="@dimen/qs_container_margin"
             android:layout_marginBottom="@dimen/close_handle_underlap" />
 
         <include layout="@layout/ambient_indication"
