+++ packages/SystemUI/src/com/android/systemui/qs/QuickStatusBarHeader.java
@@ -60,8 +60,7 @@ public class QuickStatusBarHeader extends RelativeLayout {
     protected QSTileHost mHost;
     private View mDate;
 
-	private View mHeaderView;
-	private String mHeaderColor = "#00FFFFFF";
+	
 
     public QuickStatusBarHeader(Context context, AttributeSet attrs) {
         super(context, attrs);
@@ -84,8 +83,6 @@ public class QuickStatusBarHeader extends RelativeLayout {
         float intensity = colorForeground == Color.WHITE ? 0 : 1;
         Rect tintArea = new Rect(0, 0, 0, 0);
 		
-		mHeaderView = findViewById(R.id.qs_top_background);
-		updateColor();
 
 
         BatteryMeterView battery = findViewById(R.id.battery);
@@ -109,31 +106,8 @@ public class QuickStatusBarHeader extends RelativeLayout {
         applyDarkness(R.id.qs_left_clock, tintArea, intensity, colorForeground);
 		
     }
+
 	
-	private class ColorObserver extends ContentObserver {
-        ColorObserver(Handler handler) {
-            super(handler);
-        }
-
-        void observe() {
-            getContext().getContentResolver().registerContentObserver(Settings.System
-                    .getUriFor(Settings.System.QS_HEADER_COLOR), false,
-                    this, UserHandle.USER_ALL);
-        }
-
-        @Override
-        public void onChange(boolean selfChange) {
-            updateColor();
-        }
-    }
-	
-	private void updateColor() {
-        mHeaderColor = Settings.System.getStringForUser(getContext().getContentResolver(),
-                Settings.System.QS_HEADER_COLOR,
-                UserHandle.USER_CURRENT);
-		mHeaderView.setBackgroundColor(Color.parseColor(mHeaderColor));
-    }
-
     private void applyDarkness(int id, Rect tintArea, float intensity, int color) {
         View v = findViewById(id);
         if (v instanceof DarkReceiver) {
